Add Offers Screen and Update Navigation (UI)
Added new images for offers selection and unselection.
Updated language files to include "offers" key.
Implemented OffersItemScreen and integrated it into the navigation.
Modified DashboardScreen to include OffersItemScreen in the menu.
Updated RouteHelper to include the new offers route.
Adjusted MenuScreen to include a link to the orders screen.
Modified